Parasolid file for the Ender V2 40x20mm Fan Cover
This is a subtle remix of the stock Ender V2 fan cover that has been adapted to fit a 40x20mm hot end fan, as well as other changes that make it easier to 3D print. This file requires no print supports and has been tested in PETG. Bosses exist for hot sinking M3 and M2 brass threads.